created a Connections class, refactored code to use this class instead of pg. we started running into issues, where tests would fail because too many SQL connections were open, because we weren't closing them after we were done with them. pepering the lib classes with connection.close() would hold it off for a bit, but we didn't do all of them so the issue kept coming back. The Connection class I made creates a connection, executes the query, then closes the connection. To make a query, just type Connection.exec('SELECT * FROM users;')
