+++ lib/user.rb
@@ -1,4 +1,5 @@
 require 'pg'
+require_relative 'sql_connections'
 
 class User
 =begin
@@ -19,16 +20,14 @@ to create a login method.
   class << self
 
     def create(username:, email:, password:)
-      connection = makeconnection()
       return [false, "username already taken"] if name_exists?(username)
       return [false, "email already taken"] if email_exists?(email)
-      result = connection.exec("INSERT INTO users (username, email, password) VALUES('#{username}', '#{email}', '#{password}') RETURNING *;")
+      result = Connection.exec("INSERT INTO users (username, email, password) VALUES('#{username}', '#{email}', '#{password}') RETURNING *;")
       [true, User.new(id: result[0]['id'], username: result[0]['username'], email: result[0]['email'], password: result[0]['password'])]
     end
 
     def all
-      connection = makeconnection()
-      result = connection.exec("SELECT * FROM users;")
+      result = Connection.exec("SELECT * FROM users;")
       result.map do |user|
         User.new(id: user['id'], username: user['username'], email: user['email'], password: user['password'])
       end
@@ -37,16 +36,14 @@ to create a login method.
     def login(email:, password:)
       return [false, "wrong email address"] if not email_exists?(email)
       return [false, "wrong password"] if not password_correct?(email: email, password: password)
-      connection = makeconnection()
-      result = connection.exec("SELECT * FROM users WHERE email = '#{email}';")
+      result = Connection.exec("SELECT * FROM users WHERE email = '#{email}';")
       [true, User.new(id: result[0]['id'], username: result[0]['username'], email: result[0]['email'], password: result[0]['password'])]
     end
 
     private
 
     def password_correct?(email:, password:)
-      connection = makeconnection()
-      result = connection.exec("SELECT * FROM users WHERE email='#{email}';")
+      result = Connection.exec("SELECT * FROM users WHERE email='#{email}';")
       # return BCrypt::Password.new(result[0]['password']) == password
       result[0]['password'] == password
     end
@@ -69,12 +66,5 @@ to create a login method.
       existing_emails.include?(username) ? true : false
     end
 
-    def makeconnection
-      if ENV['ENVIRONMENT'] == 'test'
-        connection = PG.connect(dbname: 'makersbnb_test')
-      else
-        connection = PG.connect(dbname: 'makersbnb')
-      end
-    end
   end
 end
